move hooks to root comp
