+++ webapp/src/components/expanded_view/remove_confirmation.tsx
@@ -18,7 +18,6 @@ type Props = {
 export const RemoveConfirmation = ({profile, onConfirm, onCancel}: Props) => {
     const {formatMessage} = useIntl();
 
-    // @ts-ignore
     const body = formatMessage({
         defaultMessage: 'Are you sure you want to remove <b>{userName}</b> from the call?',
     }, {
